文件删除及判断 部分无用文件及导入删除 在Matter表新增size字段 用户表新增isAdmin字段 管理员设置

+++ src/main/java/com/shih/icecms/controller/MatterController.java
@@ -183,6 +183,10 @@ public class MatterController {
     @ApiOperation(value = "文件删除")
     @DeleteMapping("/matter/delete")
     public ApiResult delete(@RequestParam String matterId)   {
+        // 检查是否含有文件或子文件夹
+        if (matterService.count(new LambdaQueryWrapper<Matter>().eq(Matter::getParentId,matterId))>0) {
+            return ApiResult.ERROR("文件夹内含有文件，删除失败");
+        }
         if(matterService.deleteMatter(matterId)){
             return ApiResult.SUCCESS("删除成功");
         }else{
